pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bancolombia-rag"
version = "0.1.0"
description = "Sistema RAG end-to-end para documentación pública de Bancolombia"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.13.0",
    "openai>=1.50.0",
    "chromadb>=0.5.0",
    "fastmcp>=2.0.0",
    "streamlit>=1.35.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
# line-length = 120: el proyecto incluye strings de system prompt y mensajes de error
# que son inherentemente largos; 88 fuerza roturas artificiales que dañan la legibilidad.
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
